[wip] Add script fallbacks if no fallback is found
Add the CLDR data 'script' as a fallback in case no language
fallback is found, and before settling for "en" or local language.

This brings back language_scripts dependency, but requires no
updates to that remote repo.

Please hold merging while we discuss; tests were added to show
what the process looks like. This may mitigate the issues of
semi-random language codes in OSM while still sticking to the
"proper" way of doing language fallback based *primarily* on the
language code and only if no other choice then on the script.